Added base FSM class
Added the base class for all FSMs.
Derived FSM class will be able to define all transition logic through a
set of rules which specify the origin and destination states of the
transition, and a condition function that needs to return true for the
transition to trigger.
Similarly to the FSMState class, entities will be able to use a shared
single instance of each FSM for optimal memory usage.