Added the mechanisms necessary to implement a Hierarchical FSM
Added a state class that acts as a wrapper or proxy on an FSM, so that it can be used as a "state" inside other FSMs.
Modified the State container class so that it is possible to register a state under a different class type key. This makes it possible to obtain the hierarchical proxy state instance associated with an FSM directly by providing the FSM class type.