Added stack-based state callbacks
Added new state callbacks that enable performing state logic specific to stack-based transitions.
This is done via the two new OnResume and OnPause functions, which serve as equivalent for stack-based transitions of the OnEnter and OnExit functions respectively.
If these new functions are not overridden in a state, the default transition logic will be executed by default.